Add files via upload
The goal of the project was to create the division of a 16-bit and an 8-bit binary number to obtain a 16-bit quotient. The project was done with the help of an auxiliary 16-bit binary register. The registers are 8 bits long, so each 16-bit data had two registers: LSB (Least Significant Bit) and MSB (Most Significant Bit) registers.
Other parts of the code are not included because they are part of copyright. My task was to modify only the ROM file and adapt it to my division task.
There were 20 commands provided in the Instruction Set PDF file.